requirements/base.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
httpx>=0.25.0
gql[aiohttp]>=3.4.0
tenacity>=8.2.0
PyJWT>=2.8.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
prometheus-client>=0.19.0
structlog>=23.2.0
